Isolating all power converter specs
Instead of relying on a massive test suite allow each conversion to be
tested in isolation.
