+++ app/conversions/power_converters/safe_for_method_name.rb
@@ -1,7 +1,5 @@
 PowerConverter.define_conversion_for(:safe_for_method_name) do |input|
   case input
-  when NilClass
-    nil
   when Symbol, String
     if input.present?
       input.to_s.gsub(/\W+/, '_').underscore
